Move HTMLDocument to its own module
