+++ nilearn/reporting/__init__.py
@@ -2,8 +2,9 @@
 Reporting code for nilearn
 """
 
-from .html_report import (ReportMixin, HTMLDocument, HTMLReport,
-                          set_max_img_views_before_warning)
+from .html_report import (ReportMixin, HTMLReport)
+
+from .html_doc import (HTMLDocument, set_max_img_views_before_warning)
 
 __all__ = ['ReportMixin', 'HTMLDocument', 'HTMLReport',
            'set_max_img_views_before_warning']
